clean up regPalaceView component, move deletePalaceFromBackEnd to context
